Optimize routing and file handling
Restructured the routes for product and account into separate modules (product.py and account.py respectively). Image upload functionality moved to a separate module to segregate responsibilities for cleaner, more organized code. This also results in a more intuitive directory structure. The server port has also been changed from 8090 to 8000 for standardization. Imported utility functions have also been restructured to be contained in a tool specific file (image_io.py). This enhances code readability and boosts efficient execution.
