+++ core/env.ini
@@ -1,6 +1,6 @@
 [SERVER]
 host = localhost
-port = 8090
+port = 8000
 workers = 4
 api_prefix = /api
 api_docs = /api/docs
