+++ router/Account/account.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter
+
+prefix = "/account"
+router = APIRouter(
+    prefix=prefix,
+    tags=["account"],
+)
