+++ router/Account/image.py
@@ -1,11 +1,7 @@
-from fastapi import APIRouter, UploadFile, HTTPException
-from model import image_io
+from fastapi import UploadFile
 
-prefix = "/account"
-router = APIRouter(
-    prefix=prefix,
-    tags=["account"],
-)
+from router.Account.account import router
+from utils import image_io
 
 @router.post("/{id}/upload_image",
              description="Upload image(jpeg/png) for product, max_size = 10MB",
@@ -15,3 +11,11 @@ router = APIRouter(
 async def upload_image(id, file: UploadFile):
     flag = await image_io.save_file(file, image_io.ImgSourceEnum.avatar, id)
     return {"filename": file.filename, "content_type": file.content_type, "size": file.size, "success": flag}
+
+@router.get("/{id}/get_image",
+            description="Get image(jpeg/png) for product",
+            status_code=200,
+            tags=["account", "img_get"]
+            )
+async def get_image(id):
+    pass
