+++ router/Product/product.py
@@ -1,0 +1,7 @@
+from fastapi import APIRouter
+
+prefix = "/product"
+router = APIRouter(
+    prefix=prefix,
+    tags=["product"],
+)
